async ajax messing up the file names

+++ attachment_upload.php
@@ -40,26 +40,29 @@ if(isset($_POST["doc"]) && isset($_POST["doc_id"])){
     exit;
 }
 
-if( isset($_REQUEST["attach_id"]) ){
-    $_id = $_REQUEST["attach_id"];
+if( isset($_REQUEST["attach_id"]) && isset($_REQUEST["attach_name"])){
+    $_id    = $_REQUEST["attach_id"];
+    $name   = $_REQUEST["attach_name"];
 
     $split      = strpos($_id, "_photo_") > -1 ? "_photo_" : "_audio_";
     $temp       = explode($split, $_id);
     $walk_id    = $temp[0]; 
     $local_folder   = "temp/$walk_id";
 
-    // CHECK IF WALK DATA ALREADY EXISTS, NEED TO DELETE IT TO WRITE IT AGAIN, NO OVERWRITE FEATURE?
-    $attachment     = $local_folder."/".$_id;
-    if( file_exists($attachment) ){
-        unlink($attachment);
-    }
-
     if ($_FILES["attachment"]["error"] == UPLOAD_ERR_OK){
-        $file = $_FILES["attachment"]["tmp_name"];
+        // CHECK IF WALK DATA ALREADY EXISTS, NEED TO DELETE IT TO WRITE IT AGAIN, NO OVERWRITE FEATURE?
+        $file       = $_FILES["attachment"]["tmp_name"];
+        $fd_id      = $_FILES["_id"];
+        $fd_name    = $_FILES["_name"];
+        $attachment = $local_folder."/".$fd_id;
+        if( file_exists($attachment) ){
+            unlink($attachment);
+        }
+
         // now you have access to the file being uploaded
         //perform the upload operation.
         move_uploaded_file( $file, $attachment );
-        print_r(json_encode(array($attachment . " saved to disk?!!!")));
+        print_r(json_encode(array($fd_id . " $fd_name saved to disk?!!!")));
     }else{
         print_r(json_encode(array($attachment . "upload failed")));
     }
